Pull in 0.10.0 of sc-collections, eliminating Unity dep
sc-collections uses the Unity C unit testing library. But there's no
reason for it to pull in extra dependencies for its users. Version
0.10.0 ensures that it no longer pulls in the Unity depnedency when used
as a subproject. Require that version as a minimum, and pin it since
there's no reason to update it. This is an example repository which
won't see feature updates.

This fixes issue #1.

Signed-off-by: Stephen Brennan <stephen@brennan.io>